feat: context API 상태 관리
기존 url을 활용한 상태 관리 방식을 Context API으로 변경
